Autoscheduler : refactor code
