+++ include/tiramisu/auto_scheduler/auto_scheduler.h
@@ -34,9 +34,18 @@ public:
     auto_scheduler(search_method *searcher, evaluator *eval_func,
                    tiramisu::function *fct = tiramisu::global::get_implicit_function());
                    
-    void schedule();
-                   
     void set_exec_evaluator(evaluate_by_execution *exec_evaluator) { this->exec_evaluator = exec_evaluator; }
+    
+    /**
+     * Use the search method to find a set of optimizations.
+     */
+    void find_schedule();
+    
+    /**
+     * Use the Tiramisu API to apply the schedule found by
+     * find_schedule().
+     */
+    void apply_schedule();
 };
 
 }
